+++ tests/FilterTest.php
@@ -418,6 +418,36 @@ class FilterTest extends TestCase
         $this->assertGreaterThan(0, TestModel::all()->count());
     }
 
+    /** @test */
+    public function it_should_apply_a_default_filter_value_if_nothing_in_request()
+    {
+        TestModel::create(['name' => 'John Doe']);
+        TestModel::create(['name' => 'John Deer']);
+
+        $models = $this
+            ->createQueryFromFilterRequest([])
+            ->allowedFilters(AllowedFilter::partial('name')->default('John'))
+            ->get();
+
+        $this->assertEquals(2, $models->count());
+    }
+
+    /** @test */
+    public function it_does_not_apply_default_filter_when_filter_exists_and_default_is_set()
+    {
+        TestModel::create(['name' => 'John Doe']);
+        TestModel::create(['name' => 'John Deer']);
+
+        $models = $this
+            ->createQueryFromFilterRequest([
+                'name' => 'Doe',
+            ])
+            ->allowedFilters(AllowedFilter::partial('name')->default('John'))
+            ->get();
+
+        $this->assertEquals(1, $models->count());
+    }
+
     protected function createQueryFromFilterRequest(array $filters): QueryBuilder
     {
         $request = new Request([